Move error to error.rs
